HW1: after review #2 add converter interface
